requirements.txt
# ============================
# requirements.txt
# Minimal runtime deps for Streamlit + TFLite inference
# ============================

streamlit==1.40.2
tflite-runtime==2.14.0
numpy==1.26.1
pandas==2.1.1
Pillow==10.0.1
requests>=2.31.0
